requirements.txt
fastapi
pyodbc
python-decouple
uvicorn
jinja2
requests